server/requirements/base.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
Pillow==6.2.1
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.6
